1.fix the problem that the electronic compass Mecompass is hung on the Orion mainboard 7 or 8 ports and communication will be hung dead. 2.fix the problem that the function getPointFast() in MeHumitureSensor.cpp does not normally output the value. 3.fix the problem that compile smartservo_test.ino firmware error report using the arduino1.6.5 environment with mBlock V3.4.12. 4.remove MeSuperVariable.cpp/MeSuperVariable.h. 5.fix the problem that ultrasonic module can only measure the maximum range of 375cm,and the maximum range of normal requirements is 400cm.

+++ src/MeHumitureSensor.cpp
@@ -417,7 +417,7 @@ double MeHumiture::getPointFast()
 {
   double a = 17.271;
   double b = 237.7;
-  double temp = (a * Temperature) / (b + Temperature) + log(Humidity/100);
+  double temp = (a * Temperature) / (b + Temperature) + log((double)Humidity/100);
   //double Td = (b * temp) / (a - temp);
   return ((b * temp) / (a - temp));
 }
